Organização do Código da versão 1.0.4 da AppStore
