+++ The Big Short/Model/FetchDataAPI/CurrencyData.swift
