+++ The Big Short/Model/FetchDataAPI/MultiStockData.swift
