+++ The Big Short/Model/FetchDataAPI/StockData.swift
